requirements.txt
# PDF Toolkit - Requirements
# Python 3.10+

# UI Framework
PySide6>=6.6.0

# Image Processing
Pillow>=10.0.0
img2pdf>=0.5.0
opencv-python>=4.8.0
numpy>=1.24.0

# PDF Processing
pypdf>=3.17.0
pymupdf>=1.23.0  # For PDF to Images conversion
pdf2image>=1.16.0
reportlab>=4.0.0

# OCR
pytesseract>=0.3.10

# Word Document Support
python-docx>=1.0.0
